# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: WoloApp
  version: "1.0.0"
  description: >-
    The WoloApp API facilitates the management of events and user participation
    through Create, Read, Update, and Delete operations. It allows users to
    interact with the following key functionalities such as exploring events
    through different filters or joining them.
  termsOfService: >-
    https://opensource.guide/legal/?fbclid=IwAR0NUY2SAm3x5HHkhxOmb6piB-WETsmqeifvIB5_pBrHWFhQu31oYBl7WJs
  contact:
    name: Agata Dobrzyniewicz
    email: s24305@pjwstk.edu.pl
servers:
  - url: http://localhost:8080
    description: local
paths:
...
  /events/{id}:
    get:
      tags:
        - Event
      summary: Get event by id
      description: Event must exist
      operationId: getEventById
      parameters:
        - name: id
          in: path
          description: Event id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponseDetailsDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /events/upcoming:
    get:
      tags:
        - Event
      summary: Get only upcoming events
      operationId: getUpcomingEvents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /events/users:
    get:
      tags:
        - Event
      summary: Get short information about users by shift id
      description: Shift must exist
      operationId: getUsersByShift
      parameters:
        - name: shift
          in: query
          description: shift id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserShortResponse'
  /events/users/pdf:
    get:
      tags:
        - Event
      summary: Get PDF report with list of users signed up to shifts on this event
      description: |
        Retrieves a PDF report containing short information about users assigned to a specific shift.
        The PDF report will contain details such as user names, phone numbers, and emails for the specified shift.
      operationId: getListOfUsersAsPDF
      parameters:
        - name: id
          in: query
          description: ID of the event for which to generate the PDF report
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: PDF report generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
              example: <binary PDF data>
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /events/search:
    get:
      tags:
        - Event
      summary: Filter events
      operationId: filterEvents
      parameters:
        - name: localization
          in: query
          description: Array of District names
          required: false
          schema:
            type: array
            items:
              type: string
        - name: startDate
          in: query
          description: >-
            The beginning of the time interval within which to search for an
            event
          required: false
          schema:
            type: string
            format: date
          example: '2022-01-01'
        - name: endDate
          in: query
          description: The end of the time interval within which to search for an event
          required: false
          schema:
            type: string
            format: date
          example: '2022-12-31'
        - name: categories
          in: query
          description: Array of category ids
          schema:
            type: array
            items:
              type: integer
              format: int64
        - name: organizer
          in: query
          description: Id of the organization organizing the event
          required: false
          schema:
            type: integer
            format: int64
            example: 1
        - name: ageRestriction
          in: query
          description: From what age are participants allowed
          required: false
          schema:
            type: integer
          example: 18
        - name: isPeselVerificationRequired
          in: query
          description: Whether pesel verification is required
          schema:
            type: boolean
        - name: showWithAvailableCapacity
          in: query
          description: Show only those events that are still looking for volunteers
          schema:
            type: boolean
        - name: category
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - name: verification
          in: query
          required: false
          schema:
            type: boolean
        - name: showAvailable
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /events/add:
    post:
      tags:
        - Event
      summary: Adding new event by organisation
      operationId: addEvent
      parameters:
        - name: language
          in: query
          description: Language of eventDto
          required: true
          schema:
            type: string
          example: EN
        - name: event
          in: query
          description: Event object to create
          schema:
            $ref: '#/components/schemas/EventRequestDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDto'
        required: true
      responses:
        '201':
          $ref: '#/components/responses/Created'
  /events/admin/add:
    post:
      tags:
        - Event
      summary: Adding new event by admin
      operationId: addEventByAdmin
      parameters:
        - name: language
          in: query
          description: Language of eventDto
          required: true
          schema:
            type: string
          example: EN
        - name: event
          in: query
          description: Event object to create
          schema:
            $ref: '#/components/schemas/EventRequestDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDto'
        required: true
      responses:
        '201':
          $ref: '#/components/responses/Created'
  /events/join:
    post:
      tags:
        - Event
      summary: The user join shift of the event
      operationId: joinEvent
      parameters:
        - name: shift
          in: query
          description: Id of the shift user wants to join
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /events/join/check:
    post:
      tags:
        - Event
      summary: We check whether the user has changes during the shift to which wants to be assigned
      operationId: checkJoinEvent
      parameters:
        - name: shift
          in: query
          description: Id of the shift user wants to join
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /events/refuse:
    post:
      tags:
        - Event
      summary: The user refuses to participate in the event
      operationId: refuseParticipateInEvent
      parameters:
        - name: shift
          in: query
          description: Id of the shift from which the user wants to unsubscribe
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
  /events/admin/edit/{id}:
    put:
      tags:
        - Event
      summary: Edit event by admin
      description: Only upcomming events can be changed
      operationId: editEventByAdmin
      parameters:
        - name: event
          in: query
          description: event object with changes
          schema:
            $ref: '#/components/schemas/EventRequestDto'
        - name: id
          in: path
          description: event id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: language
          in: query
          description: language of eventDto
          required: true
          schema:
            type: string
            example: EN
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDto'
        required: true
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /events/edit/{id}:
    put:
      tags:
        - Event
      summary: Edit event by organisation
      description: Only upcomming events can be changed
      operationId: editEvent
      parameters:
        - name: event
          in: query
          description: event object with changes
          schema:
            $ref: '#/components/schemas/EventRequestDto'
        - name: id
          in: path
          description: event id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: language
          in: query
          description: language of eventDto
          required: true
          schema:
            type: string
            example: EN
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestDto'
        required: true
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /events/delete/{id}:
    delete:
      tags:
        - Event
      summary: Delete event by organisation
      description: Only upcoming events can be deleted
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          description: Event id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
  /events/admin/delete/{id}:
    delete:
      tags:
        - Event
      summary: Delete event by admin
      description: Only upcoming events can be deleted
      operationId: deleteEventByAdmin
      parameters:
        - name: id
          in: path
          description: Event id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
...
  /organisations/{id}/events:
    get:
      tags:
        - Organisations
      summary: Get all events from the selected organisation
      operationId: getEventsByOrganisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by id
      description: User must exist
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /users/events/{id}:
    get:
      tags:
        - Users
      summary: Get all user events
      operationId: getUserEvents
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /users/delete/{id}:
    delete:
      tags:
        - Users
      summary: Delete user
      description: If the user has the moderator role, it cannot be deleted
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
components:
  responses:
    OK:
      description: OK
      content:
        application/hal+json:
          schema:
            type: string
            example: Ok
            enum:
              - 200 OK
    Created:
      description: Created
      content:
        application/hal+json:
          schema:
            type: string
            example: Created
            enum:
              - 201 CREATED
    NoContentResponse:
      description: No content
      content:
        application/hal+json:
          schema:
            type: string
            example: No Content
            enum:
              - 204 NO_CONTENT
    BadRequestResponse:
      description: Bad Request
      content:
        application/hal+json:
          schema:
            type: string
            example: Bad Request
            enum:
              - 400 BAD_REQUEST
    NotFoundResponse:
      description: Not Found
      content:
        application/hal+json:
          schema:
            type: string
            example: Not Found
            enum:
              - 404 NOT_FOUND
    InternalServerError:
      description:   Internal Server Error
      content:
        application/hal+json:
          schema:
            type: string
            example: Internal Server Error
            enum:
              - 500 INTERNAL_SERVER_ERROR
  schemas:
...
    EventResponseDetailsDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        namePL:
          type: string
          example: Pomoc w schronisku dla zwierząt
        nameEN:
          type: string
          example: Pomoc w schronisku dla zwierząt
        nameUA:
          type: string
          example: Допомога притулку для тварин
        nameRU:
          type: string
          example: Помощь приюту для животных
        organisationId:
          type: integer
          format: int64
          example: 1
        organisationName:
          type: string
          example: 'Compassion Paws: Animal Shelter Support Initiative'
        date:
          type: string
          format: date
        isPeselVerificationRequired:
          type: boolean
        isAgreementNeeded:
          type: boolean
        descriptionEN:
          type: string
          example: >-
            Are you passionate about making a positive impact on the lives of
            animals in need? Join us in our mission 

            to provide compassionate care and support for animals at our local
            animal shelter. 

            As an Animal Shelter Assistant, you play a crucial role in ensuring
            the well-being 

            and happiness of animals awaiting their forever homes.


            Responsibilities:

                Animal Care:
                    Feed, groom, and provide basic care for shelter animals.
                    Ensure clean and sanitary living conditions for all animals.

                Assistance with Adoptions:
                    Interact with potential adopters and provide information about available animals.
                    Assist in the adoption process, ensuring a smooth transition for both the animal and the new owner.
        descriptionPL:
          type: string
          example: >-
            Czy pasjonujesz się wywieraniem pozytywnego wpływu na życie ludzi
            zwierzęta w potrzebie? Dołącz do nas w naszej misji aby zapewnić współczującą opiekę i wsparcie zwierzętom w naszej okolicy
            Schronisko dla zwierząt.

            Jako asystent schroniska dla zwierząt odgrywasz kluczową rolę w zapewnianiu dobre samopoczucie 
            i szczęście zwierząt oczekujących na swoje stałe domy.

            Obowiązki:
            Opieka nad zwierzętami:karmić, pielęgnować i zapewniać podstawową opiekę zwierzętom ze schroniska.
            Zapewnij wszystkim zwierzętom czyste i higieniczne warunki życia.

            Pomoc przy adopcjach: wejdź w interakcję z potencjalnymi adopcjami i przekaż informacje o dostępnych zwierzętach.
            Pomagaj w procesie adopcji, zapewniając płynną zmianę zarówno zwierzęciu, jak i nowemu właścicielowi.
        descriptionUA:
          type: string
          example: >-
            Ви прагнете позитивно вплинути на життя тварини в потребі? Приєднуйтесь до нашої місії
            надавати співчутливу турботу та підтримку тваринам у нашому закладі притулок для тварин.
            Як помічник у притулку для тварин, ви відіграєте вирішальну роль у забезпеченні благополуччя
            і щастя тварин, які чекають своїх вічних домівок.
            Обов'язки: догляд за тваринами:
            Годуйте, доглядайте та забезпечуйте основний догляд за тваринами з притулку.
            Забезпечити чисті та санітарні умови проживання для всіх тварин.
            Допомога з усиновленням:
            Спілкуйтеся з потенційними усиновлювачами та надайте інформацію про доступних тварин.
            Допомога в процесі усиновлення, забезпечуючи плавний перехід як для тварини, так і для нового власника.
        descriptionRU:
          type: string
          example: >-
            Уход за животными:

                Кормление, уход и обеспечение основных потребностей животных в приюте.
                Обеспечение чистых и санитарных условий проживания для всех подопечных.

            Помощь в усыновлении:

                Общение с потенциальными усыновителями и предоставление информации о доступных животных.
                Помощь в процессе усыновления, обеспечивая плавный переход как для животного, так и для нового владельца.
        categories:
          type: array
          items:
            type: string
            example: Health
        imageUrl:
          type: string
          example: https://example.com/image.jpg
        alt:
          type: string
        shifts:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ShiftResponseDto'
        city:
          type: string
          example: New York City
    EventResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nameEN:
          type: string
          example: Animal shelter assistance
        namePL:
          type: string
        nameUA:
          type: string
        nameRU:
          type: string
        organisation:
          type: string
          example: 'Compassion Paws: Animal Shelter Support Initiative'
        isPeselVerificationRequired:
          type: boolean
        categories:
          type: array
          items:
            type: string
            example: Health
        city:
          type: string
          example: New York City
        date:
          type: string
          format: date
        imageUrl:
          type: string
          example: https://example.com/image.jpg
        alt:
          type: string
        shifts:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ShiftResponseDto'
...
    UserShortResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstName:
          maxLength: 50
          minLength: 0
          type: string
          example: John
        lastName:
          maxLength: 50
          minLength: 0
          type: string
          example: Doe
        email:
          maxLength: 50
          minLength: 0
          type: string
          example: example@mail.com
        phoneNumber:
          maxLength: 9
          minLength: 9
          pattern: '[0-9]{9}'
          type: string
...